Split upscale workspace helpers
