+++ frontend/src/components/tools/upscale/_lib/upscale-workspace-copy.ts
@@ -1,0 +1,67 @@
+export const DEFAULT_UPSCALE_COPY = {
+  disabledTitle: 'Tools are disabled',
+  disabledBody: 'Enable `FEATURES.workflows.toolsSection` to access this area.',
+  back: 'Back to Tools',
+  eyebrow: 'AI Upscale Studio',
+  title: 'Upscale images and videos',
+  subtitle: 'A clean finishing pass for stills and short clips before they go back into Library, Image, or Video.',
+  image: 'Image',
+  video: 'Video',
+  sourceTitle: 'Source',
+  upload: 'Drop or choose file',
+  urlPlaceholder: 'Paste an image or video URL',
+  uploadFailed: 'Upload failed.',
+  settingsTitle: 'Upscale recipe',
+  mediaType: 'Media type',
+  engine: 'Engine',
+  mode: 'Mode',
+  factor: 'Factor',
+  target: 'Target',
+  output: 'Format',
+  run: 'Run Upscale',
+  running: 'Upscaling...',
+  outputTitle: 'Studio preview',
+  emptyOutput: 'Add a source and run an upscale to compare the finished asset here.',
+  previewSource: 'Source',
+  previewResult: 'Result',
+  previewCompare: 'Compare',
+  previewZoom: 'Pixel zoom',
+  previewZoomFit: 'Fit',
+  save: 'Save',
+  saved: 'Saved to Library.',
+  saveFailed: 'Failed to save to Library.',
+  download: 'Download',
+  recentTitle: 'Recent upscales',
+  priceEyebrow: 'Est. price',
+  priceLoading: 'Loading price...',
+  priceUnavailable: 'Price unavailable',
+  priceReady: 'Shown before generation',
+  priceVideoLoading: 'Reading video metadata...',
+  priceVideoMissing: 'Add a readable video source to estimate.',
+  library: 'Library',
+  libraryChoose: 'Choose from Library',
+  libraryTitle: 'Choose source asset',
+  libraryBody: 'Select an image or video already saved in your Library.',
+  librarySearch: 'Search assets...',
+  librarySourcesTitle: 'Library',
+  libraryEmpty: 'No saved assets yet.',
+  libraryEmptyImages: 'No saved images yet.',
+  libraryEmptyVideos: 'No saved videos yet.',
+  libraryEmptySearch: 'No assets match this search.',
+  libraryError: 'Failed to load Library assets.',
+  libraryRefresh: 'Refresh',
+  libraryUse: 'Use',
+  libraryCount: '{count} assets',
+  libraryTabs: {
+    all: 'All assets',
+    upload: 'Uploaded',
+    generated: 'Generated',
+    character: 'Character',
+    angle: 'Angle',
+    upscale: 'Upscale',
+  },
+  authTitle: 'Sign in to upscale',
+  authBody: 'Uploads, wallet billing, and Library actions require an account.',
+  authPrimary: 'Create account',
+  authSecondary: 'Sign in',
+} as const;
